requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# HTML/CSS generation for EDM templates
jinja2>=3.1.2

# LLM API
openai>=1.0.0

# Utilities
requests>=2.31.0
base64>=0.0.0

# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
